+++ BackEnd/IdeaManager-API/src/repositories/IdeaRepository.ts
@@ -15,6 +15,11 @@ class IdeaRepository {
     this.repo = [];
   }
 
+  public findByID(ideaId: string):Idea{
+    let index = this.repo.findIndex((idea)=> idea.id === ideaId)
+    return this.repo[index]
+  }
+
   public create(data: createRepoDTO): Idea {
     let { name, purpose, description, creationDate } = data;
 
@@ -29,17 +34,18 @@ class IdeaRepository {
       return this.repo
   }
 
-  public findByName(ideaName:String):Idea{
-    let index = this.repo.findIndex((idea)=> idea.name === ideaName)
-    return this.repo[index]
-  }     
 
-  public update(ideaID:string,data:Omit<Idea,"id">): Idea{
+  public findByAlias(alias:String):Idea{
+    let index = this.repo.findIndex((idea)=> idea.alias === alias)
+    return this.repo[index]
+  }  
+
+  public update(ideaID:string,data:Omit<Idea,"id"|"creationDate">): Idea{
     
     let index = this.repo.findIndex((idea)=> idea.id === ideaID)
 
     
-    let idea = {...data,id:ideaID}
+    let idea = {...data,id:ideaID,creationDate: this.repo[index].creationDate}
 
     this.repo[index] = idea